feat: API for statistic

+++ db/migrations/20240325182229_create_paras_table.up.sql
@@ -5,6 +5,6 @@ create table if not exists Paragraphs (
     linecount int,
     wordcount int,
     charcount int,
-    avglength int,
+    avglength float,
     wordfreq varchar(100000)
 )
